REFRACTOR: toolchain-file to include.cmake + rename everything to 64/32
before on linux it was x86 or x86_64 and win32/win64 on windows, now its always "32", or "64".
we don't need a toolchain file to add things to the CMAKE_lib/include/program_PATHs .. so get rid of this.

So hereby we drop any cross compilation efforts (but make it easier to integrate one afterwards, as long as someone recompiles the dependencies for mingw and maintains that stuff)